Implement PSA FWU M update agent
Implement 'Platform Security Firmware Update for the A-profile Arm
Architecture' (DEN0118) update agent which uses 'PSA Certified
Firmware Update API 1.0' (IHI0093) as its backend. In practice this
covers the use case where the update agent is running in a Secure
Partition in the A class side and delegates the firmware update to an
M class core running Trusted Firmware-M.
The commit also contains the unit test for the update agent and its
integration to component-test.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: Iaea74be5b2ab7635942e3d3ca37f8b16cb9bc5f1