Make variable index usage robust with redundancy
If there is a single variable index if a reset happends between the
removal and rewriting of the data, the index will be lost. To solve
this issue there will be two indexes (A,B) and the new data will
always be written to the older one. This way in case of an unlucky
reset, the store will be able to keep working and only the the data
of the last write will be lost.

Change-Id: I7beaca9f3be6ac1431308e309797270134af540b
Signed-off-by: Gabor Toth <gabor.toth2@arm.com>