libpsats: fix build time include path installed
Libpsats install content refers to a build time include path, which
breaks install content portability. Any project using the installed
package on a machine where TS source is not available at the same path
as on the build machine will fail to compile, due to missing "trace.h".

As a fix use the INSTALL_INTERFACE and BUILD_INTERFACE generators and
differentiate build-time and installed include paths.

Change-Id: I3f2a550ee046e5e3687612c3580bfdcbb31b855b
Signed-off-by: Gyorgy Szing <gyorgy.szing@arm.com>