docs: Update Plantuml version
Update the version required for documentation build from 1.2017.15 to
1.2024.7.

Unfortunately Ubuntu (the reference build environment is based on)
package repository has a very old version and the decision was made to
use the latest release instead even though it needs to be installed
manually.

Update the documentation and .readthedocs.yml to ensure RTD uses the
right version.

Change-Id: Ic6222f8e79efcdc53108062941c470522459fdca
Signed-off-by: Gyorgy Szing <gyorgy.szing@arm.com>