Add Encrypted Block Store
Encryption can effectively mitigate threats associated with
Block Storage, where data at rest or data in transit protection
is not inherently provided.
This change adds an AES-CBC with ESSIV encryption component, which
can be added as an extra layer.
The encryption scheme of this version is limited and uses a
hardcoded trust anchor.

Change-Id: Idaa5c1b7cbed992dedd09b4e1cdc9eedf541bfb9
Signed-off-by: Gabor Toth <gabor.toth2@arm.com>