Prevent UEFI variable store UID collision
For each variable the SMM variable backend assigns a UID for storing the
data in PSA secure storage. The UID was generated as the hashed value of
the variable name and GUID which method had a possibility for UID
collision.
Change UID generator method for simply assigning unused UIDs in the
1..max_variables range to new variables. Also moving the variable index
UID into the upper half of the UID range.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: Iab97f97658a75d65c3860fa18782b99334d0ed1d