rse_comms: Use carveout region for pointer access protocol
The Pointer access protocol requires sending physical addresses of
payload buffers. A carveout region is introduced to exchange the
payload with the RSE. Copy the invec buffers to the region, reserve
space for outvec buffers. The physical address of the region is known.
The physical address of the buffers in the region can be calculated.

Signed-off-by: Michael Zhao <michael.zhao2@arm.com>
Change-Id: Id8c0f1454d1826b23df58550fbce9807c7f9f4ae