Fix boot-order property in SP manifests
The FF-A binding specification [1] was recently updated to require all
integer properties, including boot-order, to be represented as 32-bit
wide values.
This commit updates the SP manifests accordingly to ensure compliance
with the updated specification.

Warning: this change may break compatibility with SPMC implementations.
In particular, the OP-TEE SPMC is affected and must be updated to a
compatible version.

Signed-off-by: Gyorgy Szing <gyorgy.szing@arm.com>
Change-Id: Iec2a6545924100d3e7cf0d7349fcba4025945ed2