cpputest: fix cmake 4.0 compatibility
In certain environments (such as Yocto), the project is built using the
latest version of CMake. CMake v4.0 has dropped support for versions
earlier than v3.5, and CppUTest fails to build.
This change introduces a patch that updates the minimum required CMake
version in CppUTest.

In addition, update the component to use git am instead of apply.

Signed-off-by: Gyorgy Szing <gyorgy.szing@arm.com>
Change-Id: I68291f4573fee3c64dd19c348bd8a30aa99b0c35